permissions/api_with_restrictions/requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-filter==23.2